feat: add SDL3 software renderer fallback for X11 forwarding
GL 3.3 context creation fails under X11 forwarding (XQuartz).
Added automatic fallback to SDL_Renderer(SOFTWARE) with
imgui_impl_sdlrenderer3 backend. CPU-based pan/zoom rendering
with nearest/bilinear interpolation. --software CLI flag for
forced software mode.

+++ src/diff_engine.h
@@ -21,7 +21,7 @@ public:
 
     // Render diff into the currently-bound FBO.
     // view_w/h: FBO dimensions. img_w/h: source image dimensions.
-    void render(GLuint texA, GLuint texB,
+    void render(uintptr_t texA, uintptr_t texB,
                 const ViewportState& view,
                 int img_w,  int img_h,
                 int view_w, int view_h,
